ye olde slop cleanup
